Add test for the fmt::Debug implementation
The Debug implementation doesn't enforce any sorting of the
elements (that eventually can be done in lexicographic order at the
expenses of the memory to allocate the strings) so, tests that include
more than one element are not predictable. A shallow check could be
checking the equivalence of the characters in the string.